+++ ice/src/main/java/com/altinity/ice/cli/internal/cmd/InsertWatch.java
@@ -133,6 +133,7 @@ public class InsertWatch {
                     insertBatch.iterator().next(),
                     null,
                     true,
+                    options.useVendedCredentials(),
                     options.s3NoSignRequest(),
                     null,
                     null);